feat: split SDK into curated default and /advanced subpath entries
Introduce package `exports` subpaths so the default `@roackb2/heddle`
import is the curated rung 1-5 surface a product host needs, while the
full runtime (LLM adapters, individual tools, trace, memory, models,
awareness, agent loop, heartbeat, integrations, utilities) lives behind
`@roackb2/heddle/advanced`.

- src/index.ts is now the curated entry (rungs 1-5). Approvals moved into
  rung 4 (lifecycle) since a product host owning approval policy needs
  those types from the default entry.
- src/advanced.ts re-exports the curated surface (`export * from './index'`)
  plus building blocks and specialized runtimes — the full superset. The
  union of the two files equals the previous flat surface exactly.
- package.json `exports` maps `.` -> curated and `./advanced` -> full.
- Internal consumers that pulled advanced symbols from the root barrel
  (`@/index.js` / `../../../index.js`) now import from the advanced entry;
  curated-only consumers are unchanged.
- Top-level examples that use advanced symbols import from `../src/advanced.js`;
  the sdk/ examples stay on the curated default entry.
