+++ examples/host-events.ts
@@ -12,7 +12,7 @@
  *   HEDDLE_EXAMPLE_MODEL=gpt-5.1-codex-mini npx tsx examples/host-events.ts
  */
 
-import { AgentLoopRuntimeService, HeartbeatRunnerAgent, type AgentHeartbeatEvent, type ToolDefinition } from '../src/index.js';
+import { AgentLoopRuntimeService, HeartbeatRunnerAgent, type AgentHeartbeatEvent, type ToolDefinition } from '../src/advanced.js';
 import type { LlmAdapter, LlmResponse } from '../src/core/llm/types.js';
 
 // ---------------------------------------------------------------------------
